% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/captureAll.R
\name{captureAll}
\alias{captureAll}
\title{Send output and error stream to a character string or connection.}
\usage{
captureAll(expr, conn = NULL, markStdErr = FALSE, envir = getEvalEnv(),
  doTraceback = TRUE)
}
\arguments{
\item{expr}{a single \code{expression} to be evaluated.}

\item{conn}{currently not used. A file name or a connection, or \code{NULL} to return the output as a
character vector. If the connection is not open, it will be opened initially and closed on exit.}

\item{markStdErr}{logical, should \emph{stdout} be separated from \emph{strerr} by ASCII characters STX (002) and
ETX (#003)}

\item{envir}{the \code{environment} in which \code{expr} is to be evaluated.}

\item{doTraceback}{logical, should traceback data be constructed in case of error? If \code{FALSE} no
\code{traceback} would be possible for the captured computations. Set \code{doTraceback=FALSE} for internal commands
to avoid overwriting of most recent user's errors and \code{warnings}.}
}
\description{
Send output and error stream to a character string or connection.
}
\note{
The code is based on functions \code{capture.output} and \code{.try_silent} from package \pkg{utils}.
}
