% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doCommand.R
\name{doCommand}
\alias{doCommand}
\title{Execute R command}
\usage{
doCommand(command, ...)
}
\arguments{
\item{command}{character string with command name.}

\item{\dots}{additional arguments depending on the value of \code{command}}
}
\description{
Wrapper function to execute various R commands. Intended for internal use.
}
