% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/koBrowse.R
\name{koBrowseHere}
\alias{koBrowseHere}
\alias{koBrowseEnd}
\title{Environment browser}
\usage{
koBrowseHere()

koBrowseEnd()
}
\description{
Interrupt the execution of an expression and allow the inspection of the environment where
\code{koBrowseHere} was called from.
}
\details{
These functions are only useful when used in Komodo editor. Unlike \code{browser}, \code{koBrowseHere} does
not allow for continuing the interrupted execution. \code{koBrowseEnd} returns evalution to the global
environment.
}
