% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/koCmd.R
\name{koCmd}
\alias{koCmd}
\title{Send JavaScript command to Komodo editor.}
\usage{
koCmd(cmd, async = FALSE, host = getOption("ko.host"),
  port = getOption("ko.port"), timeout = 1, ...)
}
\arguments{
\item{cmd}{character string. a JavaScript command to be evaluated by Komodo.}

\item{async}{logical. If \code{TRUE} the command returns immediately, if \code{FALSE} it waits for the response or
until timeout.}

\item{host}{character string. Host name for the port.}

\item{port}{integer. The TCP port number.}

\item{timeout}{numeric. The timeout (in seconds) to be used for this connection.}

\item{\dots}{additional arguments. Currently ignored.}
}
\description{
Send JavaScript command to Komodo editor.
}
\note{
The JavaScript \code{window} object has to be referenced explicitely via
\code{_W}.
}
