% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/koDebug.R
\name{koDebug}
\alias{koDebug}
\title{Browsing after an error}
\usage{
koDebug(FUN, refresh = TRUE)
}
\arguments{
\item{FUN}{the function to modify}

\item{refresh}{logical, if \code{TRUE} (default), R Object Browser widget is refreshed.}
}
\value{
a new function, being \code{FUN} wrapped in a debugging code that executes \code{koBrowseHere} after an error.
}
\description{
This function allows for debugging (browsing) after an error occurred in the environment
(frame) that caused the error. Similar to \code{recover} with only the innermost frame available.
It does not set a debug flag like \code{debug} does, instead returns a modified function \code{FUN}.
}
