% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/progressBar.R
\encoding{utf-8}
\name{koProgressBar}
\alias{koProgressBar}
\alias{setKoProgressBar}
\alias{getKoProgressBar}
\alias{close.koProgressBar}
\title{Progress Bars in Komodo}
\usage{
koProgressBar(title = "R progress bar", label = "", min = 0,
  max = 1, initial = 0, width = 300L)

getKoProgressBar(pb)

setKoProgressBar(pb, value, title = NULL, label = NULL)

\method{close}{koProgressBar}(con, ...)
}
\arguments{
\item{title, label}{character strings, giving the window title and the label
on the dialog box respectively.}

\item{min, max}{(finite) numeric values for the extremes of the progress bar.}

\item{initial, value}{initial or new value for the progress bar.}

\item{width}{the width of the progress bar in pixels.}

\item{pb, con}{an object of class "winProgressBar".}

\item{\dots}{for consistency with the generic.}
}
\value{
See \code{winProgressBar}.
}
\description{
Put up a Komodo progress bar widget.
}
\details{
\code{koProgressBar} will disenc2utf8play a progress bar as a dialog box inside Komodo.
The usage is equivalent to \code{winProgressBar} or \code{tkProgressBar}.
}
\author{
Kamil Barto\enc{ń}{n}
}
