% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getHelpURL.R, R/internal.R, R/utils.R
\name{getHelpURL}
\alias{getHelpURL}
\alias{kor-internal}
\alias{koPager}
\alias{koBrowser}
\alias{isInstalledPkg}
\title{Internal package functions}
\usage{
getHelpURL(topic, package = NULL)

koPager(files, header, title, delete.file)

koBrowser(url)

isInstalledPkg(pkgName, json = TRUE)
}
\arguments{
\item{topic, package}{see \code{help}}

\item{files}{name of the file to display}

\item{header}{unused argument}

\item{title}{document title}

\item{delete.file}{logical. Should the file be deleted afterwards?}

\item{url}{the URL to open.}

\item{pkgName}{the name of a package.}

\item{json}{logical, if \code{TRUE} prints result in JSON format.}
}
\description{
\code{getHelpURL} gets the URL of an R help page.

Various internal function. They are not to be called directly.

`koBrowser` opens an URL in Komodo help browser

\code{isInstalledPkg} checks if a package is installed.
}
