% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parseText.R
\name{parseText}
\alias{parseText}
\title{Parse expressions}
\usage{
parseText(text)
}
\arguments{
\item{text}{character vector. The text to parse. Elements are treated as if they were lines of a file.}
}
\description{
Parses a string and returns an unevaluated expression, an \code{"error"} object in case of
invalid code, or \code{NA} if the (most recent) expression is incomplete.
}
