% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pkgman.R
\name{pkgMan}
\alias{pkgMan}
\alias{getRepositories}
\alias{pkgManGetPackages}
\alias{pkgManGetRepositories}
\alias{pkgManGetAvailable}
\alias{pkgManGetMirrors}
\alias{pkgManSetCRANMirror}
\alias{pkgManGetDescription}
\alias{pkgManGetInstalled}
\alias{getAvailable}
\alias{pkgManRemovePackage}
\alias{pkgManLoadPackage}
\alias{pkgManDetachPackage}
\alias{pkgManGetUpdateable}
\title{Package manager supporting routines}
\usage{
getRepositories()

pkgManGetPackages(pkgName, type = "installed")

pkgManGetRepositories()

pkgManGetAvailable(page = "next", pattern = "", ilen = 50,
  col = c("Package", "Version", "InstalledVersion", "Status",
  "RepositoryName", "Priority"), reload = FALSE)

pkgManGetMirrors()

pkgManSetCRANMirror(url)

pkgManGetDescription(pkgName, print = TRUE, omitFields = NULL,
  revDep = TRUE, ...)

pkgManGetInstalled()

getAvailable()

pkgManRemovePackage(pkgName)

pkgManLoadPackage(pkgName)

pkgManDetachPackage(pkgName)

pkgManGetUpdateable()
}
\arguments{
\item{pkgName}{character vector of package names}

\item{type}{either \code{"installed"} or \code{"available"}}

\item{page}{character string. Either "prev", "next" (default), "first", "last" or "current".}

\item{pattern}{an optional regular expression. Only names matching pattern are returned.}

\item{ilen}{numeric, number of records per page.}

\item{col}{vector of character strings. Column names to be returned.}

\item{reload}{Should new data be produced if a cached copy exists?}

\item{url}{character string. CRAN mirror URL.}

\item{print}{logical, if \code{TRUE} the result is printed in DCF format.}

\item{omitFields}{a character vector giving the tags of fields to be omitted.}

\item{revDep}{logical, should reverse dependencies (only within installed
packages) be added to standard fields?}

\item{\dots}{optional, additional arguments passed to
\code{pkgReverseDependencies}. Possible arguments are:
\code{revFields} (a character vector of fields to examine for reverse
dependencies, must be one/some of \code{"Depends"}, \code{"Imports"}, \code{"Suggests"}
and \code{"Enhances"}.}

\item{installDeps}{logical. Install package dependencies?}

\item{ask}{if \code{TRUE}, asks for confirmation first.}
}
\value{
`pkgMan*` functions return `NULL` invisibly, with the side effect
   the result being printed serialized in JSON format.
}
\description{
These functions are used by the R package manager.
}
