% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Rmd2html.R
\name{rmdToHtml}
\alias{rmdToHtml}
\title{Convert Rmarkdown file to HTML}
\usage{
rmdToHtml(file, pandoc = NULL, verbose = FALSE, ...)
}
\arguments{
\item{file}{the path to the Rmarkdown input file.}

\item{pandoc}{character string. Path to 'pandoc' executable or `NULL` in which case `pandoc` is searched
for on the system path.}

\item{verbose}{if `TRUE`, messages from `knitr` and `pandoc` are printed.}

\item{\dots}{additional arguments. Currently ignored.}
}
\value{
The path to the generated HTML file is returned.
}
\description{
Convert Rmarkdown file to HTML
}
