% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rmElement.R
\name{rmElement}
\alias{rmElement}
\title{Remove element}
\usage{
rmElement(expr, envir)
}
\arguments{
\item{expr}{an object to be set to \code{NULL}.}

\item{envir}{the environment in which \code{expr} is to be evaluated.}
}
\description{
\code{rmElement} is a helper function to remove object elements by assigning a \code{NULL} value to them.
}
\note{
This may also result in removal of the parent element, if the parent element becomes \code{NULL}.
}
