% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rserver.R
\encoding{utf-8}
\name{rserver}
\alias{rserver}
\alias{startServer}
\alias{enumServers}
\alias{stopAllServers}
\alias{enumConnections}
\alias{stopAllConnections}
\title{Communication between R and Komodo}
\usage{
startServer(port)

enumServers()

stopAllServers()

enumConnections()

stopAllConnections()
}
\arguments{
\item{port}{integer. The port number.}
}
\value{
The result is an object with two components \code{"result"} and \code{"message"}.
The "message" must be one of: "more" (incomplete code, waiting for continuation), "parse-error" or "done".
In the "result", element 'stdout' and 'stdin' streams are delimited by ASCII characters #003 and #002.
}
\description{
Communication between R and a client through a socket connection.
}
\details{
Result is evaluated in R and sent back serialized in JSON format.
The client should format the data in a following way:
\itemize{
\item escape newline, carriage returns, formfeeds and backslashes with a backslash
\item if the first character is ASCII #001 then the next character is interpreted as evaluation mode specifier
\item command ends with a newline.
}
}
\section{Functions}{
\itemize{
\item \code{enumServers}: enumerate running servers

\item \code{stopAllServers}: stop all running servers

\item \code{enumConnections}: enumerate all open connections

\item \code{stopAllConnections}: close all open connections
}}

\note{
Multiple servers can be started (on different ports), and each can
simultanously accept multiple connections.
}
\author{
Kamil Barto\enc{ń}{n}
}
