% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sourceTemp.R
\name{sourceTemp}
\alias{sourceTemp}
\title{Read R code from a temporary file and delelte it afterwards}
\usage{
sourceTemp(file, local = FALSE, encoding = "utf-8", ...)
}
\arguments{
\item{file}{a character string giving the pathname of the file to read from.}

\item{local}{\code{TRUE}, \code{FALSE} or an environment, determining where the parsed expressions are evaluated.}

\item{encoding}{character vector. The encoding(s) to be assumed when file is a character string. Unlike in}

\item{\dots}{other arguments passed to \code{source}.}
}
\description{
\code{sourceTemp} differs from \code{source} only in that it deletes the \code{file} after it has been evaluated.
}
