% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/startup.R
\name{startup}
\alias{startup}
\title{Startup tasks}
\usage{
startup(verbose = FALSE)
}
\arguments{
\item{verbose}{logical. Should verbose messages be printed?}
}
\description{
This function initializes connection with Komodo and is not intended to be called by the user.
}
\keyword{internal}
