% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stringize.R
\name{stringize}
\alias{stringize}
\title{Serialize R objects}
\usage{
stringize(x)
}
\arguments{
\item{x}{a `list`}
}
\description{
A simple JSON-serializer for lists containing character strings.
}
\note{
All values are taken as strings.
}
