% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getHelpURL.R, R/internal.R, R/qParse.R,
%   R/utils.R
\name{kor-internal}
\alias{kor-internal}
\alias{getHelpURL}
\alias{koPager}
\alias{koBrowser}
\alias{qParse}
\alias{isInstalledPkg}
\alias{.instapack}
\title{Internal package functions}
\usage{
getHelpURL(topic, package = NULL)

koPager(files, header, title, delete.file)

koBrowser(url)

qParse(filename, encoding = "UTF-8")

isInstalledPkg(pkgName, json = TRUE)

.instapack()
}
\arguments{
\item{topic, package}{see \code{help}}

\item{files}{name of the file to display}

\item{header}{unused argument}

\item{title}{document title}

\item{delete.file}{logical. Should the file be deleted afterwards?}

\item{url}{the URL to open.}

\item{filename}{name of the file to parse}

\item{encoding}{The name of the encoding to be assumed. See \sQuote{Encoding} section in \code{file}.}

\item{pkgName}{the name of the package.}

\item{json}{logical, if \code{TRUE} prints result in JSON format.}
}
\description{
\code{getHelpURL} gets the URL of an R help page.

Various internal function. They are not to be called directly.

`koBrowser` opens an URL in Komodo help browser

\code{qParse} parses R code in a file and returns error message in case of error or an empty string. Used by Komodo's R-linter.

\code{isInstalledPkg} checks if a package is installed.

\code{.instapack} lists the installed, but not loaded, packages.
}
