% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/objList.R, R/objSearch.R
\name{objList}
\alias{objList}
\alias{write.objList}
\alias{objSearch}
\title{Functions used by R object browser}
\usage{
objList(id = "default", envir = getEvalEnv(), expr = NULL,
  all.names = FALSE, pattern = NULL, group = NULL, all.info = FALSE,
  compare = TRUE, ...)

write.objList(x, file = "", sep = " ", eol = "\\n", header = !attr(x,
  "all.info"), fileEncoding = "", ...)

objSearch(sep = "\\t", compare = TRUE)
}
\arguments{
\item{id}{character string.}

\item{envir}{which environment to use in listing the objects. Defaults to the current evaluation environment.}

\item{expr}{optional, the object of which elements should be listed.}

\item{all.names}{logical. If \code{FALSE}, names which begin with a \code{.} are omitted.}

\item{pattern}{optional, a regular expression. Only names matching pattern are returned.}

\item{group}{optional, name of the group of object classes. If provided, only object of that group are returned.}

\item{all.info}{if \code{TRUE}, the first column holds the environment name.}

\item{compare}{if \code{TRUE} and the result is identical to the previous one with the same \code{id}, an empty list is returned.}

\item{\dots}{additional arguments. Ignored.}

\item{x}{an object of class \code{objList}}

\item{file, sep, eol, fileEncoding}{see \code{write.table}}

\item{header}{logical. Should a header "Env=environment Obj=object" be printed?}
}
\description{
Internal functions for Komodo's R object browser.
}
\examples{
 objList()
 
if(require(datasets)) {
   write.objList(objList(expr=npk, compare=FALSE))
   objList(expr=Loblolly)
   objList(expr=Loblolly) ## an empty list (compare is TRUE)
}
}
