% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/objSearch.R, R/objls.R
\name{objSearch}
\alias{objSearch}
\alias{objls}
\alias{write.objList}
\title{Functions used by R object browser}
\usage{
objSearch(fsep = ";", rsep = "\\t")

objls(expr = NULL, envir = getEvalEnv(), all.names = FALSE,
  attrib = FALSE, ...)

write.objList(x, file = "", sep = " ", eol = "\\n", fileEncoding = "",
  ...)
}
\arguments{
\item{fsep, rsep}{character strings for field and record separator, respectively.}

\item{expr}{optional, the object of which elements should be listed.}

\item{envir}{which environment to use in listing the objects. Defaults to the current evaluation environment.}

\item{all.names}{logical. If \code{FALSE}, names which begin with a \code{.} are omitted.}

\item{attrib}{logical. If \code{TRUE} and \code{expr} is not missing, lists object's \code{attributes} after object's elements.}

\item{\dots}{additional arguments. Ignored.}

\item{x}{an object of class \code{objList}}

\item{file, sep, eol, fileEncoding}{see \code{write.table}}
}
\description{
Internal functions for Komodo's R object browser.
}
\examples{
 objls()
 
if(require(datasets)) {
   write.objList(objls(npk))
   objls(Loblolly)

}
}
