% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tempEnv.R
\encoding{utf-8}
\name{assignTemp}
\alias{assignTemp}
\alias{getTemp}
\alias{rmTemp}
\alias{existsTemp}
\alias{rmTemp}
\alias{getTemp}
\alias{getEvalEnv}
\alias{setEvalEnv}
\title{temporary environment}
\usage{
assignTemp(x, value)

existsTemp(x, mode = "any")

rmTemp(x)

getTemp(x, default = NULL, mode = "any", item = NULL)

getEvalEnv()

setEvalEnv(envir = .GlobalEnv)
}
\arguments{
\item{x, item}{a variable or list item name, given as a character string.}

\item{value}{a value to be assigned to x.}

\item{mode}{the mode or type of object sought.}

\item{envir}{the \code{environment} to use for evaluation of commands in Komodo.}
}
\description{
Maintain temporary environment. Set evaluation environment.
}
\keyword{internal}
